requirements.txt
pillow
numpy==1.26.4
torch==2.2.2+cu118
torchvision==0.17.2+cu118
scikit-learn==1.6.0
tqdm
matplotlib
pandas
flask